tuned webpack eslint pack.json and split code on modules
