+++ public/index.html
@@ -35,7 +35,7 @@
 
     <!-- Контент -->
   </main>
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 
 </html>
