+++ src/components/load-more-button.js
@@ -1,0 +1,3 @@
+const createLoadMoreButtonTemplate = () => `<button class="load-more" type="button">load more</button>`;
+
+export { createLoadMoreButtonTemplate };
